move accordion methods to a custom hook
